Cache GL Display, Context and Renderer
Instead of recreating these every time, we should make sure they're
stored in a single place.  The GL Display and Renderer can be reused
everywhere, but the contexts are specific to a config which again is
dependent on the format that we render with it.  This format

Note also that the `Renderer` can only be created with a currented
`Context` which requires a window/`Surface` (even though creating
shaders and uploading a VBO has nothing to do with WSI...), hence it is
created lazily when the first window is being rendered.